thecentralnewsnetwork.com
We Pledge that- We will not Tarnish our Values of Mainstream Journalism as a Platform for Honesty, Integrity, Accountability, Intelligence and Intensive Thinking. We will always be Steadfast in Revealing the Truth while Protecting the Secrets Surrounding our State. We will Respect all the Rules of our State, Mainstream colleagues and our State Press Institute's and Highlight all Developments and Progress of the State to the Whole World Including the Asian Continent. Journalism is a Noble Profession.